Cleanup of Attribute, remove unused constructor (only used in test), move methods used in only one place inline.
